Reconfigure layout (#606)
* Migrate layout to server

* fix a bunch of crap (#607)

* upgrade to next 15

* commented out session storage

* commented out crasher

* Moved some files and implemented naming conventions

* moved some css around to fix styling

* more refactor

* accidently removed dynamic export

* trying to disable vercel cache

---------
