+++ next/sentry.client.config.js
@@ -9,32 +9,51 @@ const tracingOrigins = [
   'empowerplant.io',
   'run.app',
   'appspot.com',
-  'vercel.app',
   /^\//,
+  window.location.host,
 ];
 
 Sentry.init({
   dsn: process.env.NEXT_PUBLIC_DSN,
-
-  // Add optional integrations for additional features
-  integrations: [
-    Sentry.replayIntegration(),
-    Sentry.browserTracingIntegration(),
-  ],
-
-  // Define how likely traces are sampled. Adjust this value in production, or use tracesSampler for greater control.
-  tracesSampleRate: 1,
+  tracesSampleRate: 1.0,
   tracePropagationTargets: tracingOrigins,
-
-  // Define how likely Replay events are sampled.
-  // This sets the sample rate to be 10%. You may want this to be 100% while
-  // in development and sample at a lower rate in production
-  replaysSessionSampleRate: 0.1,
-
-  // Define how likely Replay events are sampled when an error occurs.
-  replaysOnErrorSampleRate: 1.0,
   profilesSampleRate: 1.0,
-
-  // Setting this option to true will print useful information to the console while you're setting up Sentry.
+  replaysSessionSampleRate: 1.0,
   debug: false,
+  integrations: [
+    Sentry.replayIntegration({
+      // Additional configuration goes in here
+      // replaysSessionSampleRate and replaysOnErrorSampleRate is now a top-level SDK option
+      blockAllMedia: false,
+      // https://docs.sentry.io/platforms/javascript/session-replay/configuration/#network-details
+      networkDetailAllowUrls: ['/checkout', '/products', '/products-sc'],
+      unmask: ['.sentry-unmask'],
+    }),
+  ],
+  beforeSend(event, hint) {
+    // Parse from tags because src/index.js already set it there. Once there are React route changes, it is no longer in the URL bar
+    let se;
+    Sentry.withScope(function (scope) {
+      se = scope._tags.se;
+    });
+
+    if (se) {
+      const seTdaPrefixRegex = /[^-]+-tda-[^-]+-/;
+      let seFingerprint = se;
+      let prefix = seTdaPrefixRegex.exec(se);
+      if (prefix) {
+        // Now that TDA puts platform/browser and test path into SE tag we want to prevent
+        // creating separate issues for those. See https://github.com/sentry-demos/empower/pull/332
+        seFingerprint = prefix[0];
+      }
+
+      event.fingerprint = ['{{ default }}', seFingerprint];
+    }
+
+    if (event.exception) {
+      sessionStorage.setItem('lastErrorEventId', event.event_id);
+    }
+
+    return event;
+  },
 });
