+++ next/sentry.server.config.js
@@ -10,6 +10,8 @@ Sentry.init({
   // Define how likely traces are sampled. Adjust this value in production, or use tracesSampler for greater control.
   tracesSampleRate: 1,
 
+  integrations: [Sentry.prismaIntegration()],
+
   // Setting this option to true will print useful information to the console while you're setting up Sentry.
   debug: false,
 });
