+++ next/src/app/products-sc/page.jsx
@@ -1,0 +1,15 @@
+import ThreeDotLoader from '/src/ui/ThreeDotLoader';
+import ProductCatalog from '/src/ui/ProductCatalog';
+import { Suspense } from 'react';
+
+export const dynamic = 'force-dynamic';
+export const fetchCache = 'force-no-store';
+
+
+export default function Products() {
+  return (
+    <Suspense fallback={<ThreeDotLoader />}>
+      <ProductCatalog />
+    </Suspense>
+  );
+}
