+++ next/src/app/products/page.jsx
@@ -4,7 +4,7 @@ import * as Sentry from '@sentry/nextjs';
 import { connect } from 'react-redux';
 import { setProducts, addProduct } from '/src/actions';
 import measureRequestDuration from '/src/utils/measureRequestDuration';
-import Loader from 'react-loader-spinner';
+import ThreeDotLoader from '/src/ui/ThreeDotLoader'
 import ProductCard from '/src/components/ProductCard';
 import { useState, useEffect } from 'react';
 import { useSearchParams } from 'next/navigation';
@@ -13,8 +13,7 @@ import {
   determineBackendUrl,
 } from '/src/utils/backendrouter';
 
-function Products(props) {
-  console.log(props);
+function Products() {
   const { backend,
     frontendSlowdown,
     se,
@@ -39,10 +38,6 @@ function Products(props) {
   }
 
   function fetchUncompressedAsset() {
-    let se; // `se` is automatically added to all fetch requests, but need to do manually for script tags
-    Sentry.withScope(function (scope) {
-      se = scope._tags.se;
-    });
 
     let uc_small_script = document.createElement('script');
     uc_small_script.async = false;
@@ -149,7 +144,6 @@ function Products(props) {
 
     getProducts(frontendSlowdown);
   }, []);
-  console.log(products);
   return products.length > 0 ? (
     <div>
       <ul className="products-list">
@@ -182,9 +176,7 @@ function Products(props) {
       </ul>
     </div>
   ) : (
-    <div className="loader-container">
-      <Loader type="ThreeDots" color="#f6cfb2" height={150} width={150} />
-    </div>
+    <ThreeDotLoader />
   );
 }
