+++ next/src/ui/HomeContent.jsx
@@ -1,16 +1,14 @@
-
 'use client'
-
-import React from "react";
 
 import * as Sentry from '@sentry/nextjs';
 import { createStore, applyMiddleware, compose } from 'redux';
 import logger from 'redux-logger';
 import rootReducer from '/src/reducers';
+
+import Footer from "/src/components/Footer";
 import { Provider } from 'react-redux';
 import ScrollToTop from '/src/components/ScrollToTop';
 import Nav from '/src/components/Nav';
-
 
 import '/src/styles/index.css';
 import '/src/styles/footer.css';
@@ -21,8 +19,6 @@ import '/src/styles/cart.css';
 import '/src/styles/checkout.css';
 import '/src/styles/complete.css';
 import '/src/styles/product.css';
-import Footer from "/src/components/Footer";
-import SentryQueryInitializer from "../ui/sentry-query-initializer";
 
 
 const sentryReduxEnhancer = Sentry.createReduxEnhancer({});
@@ -33,22 +29,17 @@ const store = createStore(
 );
 
 
-export default function RootLayout({
-  // Layouts must accept a children prop.
-  // This will be populated with nested layouts or pages
-  children,
-}) {
+export default function HomeContent({children}) {
   return (
-      <html lang="en">
-        <body id="body-container">
-        <SentryQueryInitializer />
-          <Provider store={store}>
-            <ScrollToTop />
-            <Nav />
-            {children}
-          </Provider>
-          <Footer />
-        </body>
-      </html>
+    <>
+      <Provider store={store}>
+        <ScrollToTop />
+        <Nav />
+        <div id="body-container">
+        {children}
+        </div>
+      </Provider>
+      <Footer />
+    </>
   )
 }
