+++ package.json
@@ -1,0 +1,5 @@
+{
+  "dependencies": {
+    "react-loader-spinner": "^6.1.6"
+  }
+}
